Hackathon_0/Gold/AI_Employee_Vault/mcp_servers/social_media/requirements.txt
# Social Media MCP Server Dependencies
mcp>=1.0.0
requests>=2.31.0
python-dotenv>=1.0.0
facebook-sdk>=3.1.0
tweepy>=4.14.0
